% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_panel.R
\name{panel_collect}
\alias{panel_collect}
\title{Collect a panel, from wide to long}
\usage{
panel_collect(data, cross.section = NULL, cross.section.columns = NULL,
  time.variable = NULL, time.variable.columns = NULL)
}
\arguments{
\item{data}{The panel to transform}

\item{cross.section}{The name of the transformed cross sectional variable supply as chracter.}

\item{cross.section.columns}{The names of the columns indicating cross sections to collect.}

\item{time.variable}{The name of the transformed time variable supply as character.}

\item{time.variable.columns}{The names of the columns indicating time variables to collect.}
}
\value{
A collected data.table, with new columns constructed by collecting from the wide format.
}
\description{
Transforms cross sectional/time dummies to unified variables
}
\details{
For time variables named like "Time_Var_i" with arbitrary i, the program will check that all time variables are named using this convention, and strip this convention
}
\examples{

\dontrun{

#'    x_1 <- rnorm( 10000 )
   x_2 <- rnorm( 10000 ) + 0.5 * x_1
   x_3 <- rnorm( 10000 ) + 0.1 * x_1
   x_4 <- rnorm( 10000 ) - 0.7 * x_3
   x_5 <- rnorm( 10000 ) + 0.5 * x_4
   x_6 <- rnorm( 10000 )
   x_7 <- rnorm( 10000 )
   x_8 <- rnorm( 10000 )
   x_9 <- rnorm( 10000 )
   cross_levels <- c( "AT", "DE", "NO", "SWE", "GB", "CZE", "ES", "IS", "PL", "HU")
   time <- seq(1:1000)
   time <- rep(time, 10)
   geo_list <- list()
   for(i in 1:length(cross_levels))
   {  geo <- rep( cross_levels[i],
                 1000 )
                   geo_list[[i]] <- geo }
   geo <- unlist(geo_list)
   geo <- as.data.frame(geo)

   correl_data <-  do.call ( cbind, list( time, x_1, x_2,
                                           x_3, x_4, x_5,
                                           x_6, x_7, x_8,
                                           x_9))
   correl_data <- as.data.frame(correl_data)
   correl_data <- cbind( geo,
                         correl_data)

                         names(correl_data) <- c("geo", "time", "x_1",
                                                 "x_2", "x_3",  "x_4",
                                                 "x_5", "x_6",  "x_7",
                                                 "x_8", "x_9")

 X <- correl_data

panel_collect( X,
               cross.section = "geo",
               cross.section.columns = c( "SWE", "AT", "NO" ))

panel_collect( X,
               cross.section = "geo",
               cross.section.columns = c( "AT", "CZE", "DE",
                                          "ES", "GB",  "HU",
                                         "IS", "NO",  "PL",
                                          "SWE"),
               time.variable = "time",
               time.variable.columns = c( "Time_Var_1",
                                          "Time_Var_2",
                                          "Time_Var_3",
                                          "Time_Var_4",
                                          "Time_Var_5" ))

panel_collect( X,
               time.variable = "time",
               time.variable.columns = c( "Time_Var_1",
                                          "Time_Var_2",
                                          "Time_Var_3",
                                          "Time_Var_4",
                                          "Time_Var_5" ))

}


}
